pyproject.toml
[tool.poetry]
name = "web-game"
version = "0.1.0"
description = ""
authors = ["Tomasz Kordiak <dotomika@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask-restful = "^0.3.10"
pydantic = { extras = ["dotenv"], version = "^2.1.1" }
psycopg2-binary = "^2.9.7"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
isort = "^5.12.0"
coverage = "^7.3.0"
pytest = "^7.4.0"
types-psycopg2 = "^2.9.21.11"

[tool.taskipy.settings]
use_vars = true

[tool.taskipy.variables]
root_directory = "app"
tests_directory = "tests"
additional_directories = "database"

[tool.taskipy.tasks]
formatting = "task formatting_isort && task formatting_black"
formatting_black = "black {root_directory} {additional_directories} {tests_directory}"
formatting_isort = "isort {root_directory} {additional_directories} {tests_directory}"
lint = "task lint_mypy && task lint_flake8"
lint_flake8 = "flake8 {root_directory} {additional_directories} && echo All is ok!"
lint_mypy = "mypy {root_directory} {additional_directories}"
test = "coverage run -m pytest"

[tool.isort]
profile = "black"

[tool.black]
line-length = 128
